feat: ajustes no retorno da Guia de Retirada e Passando CancellationToken para alguns metodos

+++ WebZi.Plataform.Domain/ViewModel/Atendimento/AtualizarStatusLEParameters.cs
@@ -5,8 +5,8 @@ namespace WebZi.Plataform.Domain.ViewModel.Atendimento;
 public class AtualizarStatusLEParameters
 {
     [Required(ErrorMessage = "Propriedade obrigatória")]
-    public long IdentificadorProcesso { get; set; }
+    public int IdentificadorProcesso { get; set; }
 
     [Required(ErrorMessage = "Propriedade obrigatória")]
-    public long IdentificadorUsuario { get; set; }
+    public int IdentificadorUsuario { get; set; }
 }
